lab3/app/requirements.txt
streamlit==1.32.0
tensorflow-cpu==2.12.0
Pillow==10.0.0
numpy==1.23.5 